rio: Reorganize RIO server code for DS operations
DS operations will start differing from MDS operations in
subtle and not so subtle ways, as we start dealing with
dirty inodes and such.

This commit segregates the notion of a common server library
and organizes the code to be able to start adding different
routines for the DS and MDS bricks as appropriate.

Change-Id: I22c585e81e3e339984e928c50f012631d2206fad
Signed-off-by: ShyamsundarR <srangana@redhat.com>