rio: Add ability to handle dirty inodes in MDS FOPs
Dirty inodes are those that have an active reference on it
which may hence change metadata on the DS for the same.

As a result for such inodes/GFIDs we need to fetch additional
iatt information from the DS before returning the iatt for any
FOP that has it in its return parameter.

This patch introduces how that is handled from the MDS instance
of rio. It treats all inodes as dirty, as we still need the patch
that saves time on the MDS, before we can consider the inode as
being clean.

This patch further adds stat and fstat FOPs to the code, and thus
enables clients to also read data that was not possible till the
last patch.

Change-Id: I5add077bd8a5d2a762ba48b6c91e1525b755b171
Signed-off-by: ShyamsundarR <srangana@redhat.com>