fuse: add --lru-limit option
The inode LRU mechanism is moot in fuse xlator (ie. there is no
limit for the LRU list), as fuse inodes are referenced from
kernel context, and thus they can only be dropped on request of
the kernel. This might results in a high number of passive
inodes which are useless for the glusterfs client, causing a
significant memory overhead.

This change tries to remedy this by extending the LRU semantics
and allowing to set a finite limit on the fuse inode LRU.

Technically, we make inode pruning dynamic: instead of just
moving over-lru-limit inodes to the purge list (the "death row"
of the inodes), one can specify the desired actions for these
inodes in three stages:
- prune_do() is the core operation done under table lock (an
  example of it being the moving of excess inodes to purge
  list);
- it is preceded by prune_init();
- it is followed by prune_fini();
-- the latter two are run without the lock. The op table that
carries these ops gets the type prune_ops_t.

An inode table can be now created with specific alternative
prune_ops_t, which will then be applied when pruning the table.
For special use cases inode_unref() also gets a variant that takes
a prune_op_t, which will then override the prune_op_t of the table.

We apply these for fuse, where the inode table is crated with
such a prune_ops_t so that FUSE_NOTIFY_INVAL_ENTRY is sent for
the excess inodes. This has the effect of the kernel sending
FUSE_FORGET to glusterfs on behalf of those inodes and thus
eventually they are dropped. The FORGET handler has to make an
exception in its usage of inode_unref() to apply standard
prune_ops_t to achieve actual reaping of inodes (as the table's
prune_ops_t would result in an infinite cycle of sending
FUSE_NOTIFY_INVAL_ENTRY, FUSE_FORGET, FUSE_NOTIFY_INVAL_ENTRY,
...).

The default behavior remains unchanged, but glusterfs /
mount.glusterfs learns of the --lru-limit=N / -olru-limit=N
option through which N shall be set as the lru_limit of the fuse
inode table, making the above behavior effective.

Change-Id: Ifc9e111bbb914f95e26f39a586be2471b215f525
Signed-off-by: Csaba Henk <csaba@redhat.com>